+++ app/src/main/cpp/Objects/System/SettingManager/SettingManager.h
@@ -12,10 +12,12 @@ using namespace ASEngine;
 class SettingManager : public GameObject {
 
 public:
-	bool audio = true;
+	bool audio = false;
 	bool colorBlindAccessibility = false;
-
 private:
+	OboeSinePlayer sinePlayer;
+	bool isPlaying = false;
+
 	void onCreate();
 
 	void onUpdate(float delta);
